Added policy for Equiped, + reworked front.

+++ app/Http/Controllers/EquipedController.php
@@ -23,25 +23,27 @@ class EquipedController extends Controller
      */
     public function create()
     {
-        return view("equiped.create", ["equipment" => Equipment::where('is_available', true)->get()]);
+        $this->authorize("create", Equiped::class);
+        return view("equiped.create", ["equipment" => Equipment::where("is_available", true)->get()]);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request, )
+    public function store(Request $request)
     {
-        $formFields = $request->validate([
-            'equipment_id' => 'required',
+        $this->authorize("create", Equiped::class);
+        $form = $request->validate([
+            "equipment_id" => "required",
         ]);
 
-        if (Session::has('room_id')) {
-            $formFields["room_id"] = Session::get('room.id');
-            Equiped::create($formFields);
-            Session::remove('room_id');
+        if (Session::has("room_id")) {
+            $form["room_id"] = Session::get("room.id");
+            Equiped::create($form);
+            Session::remove("room_id");
             return redirect("/")->with("success", "You have connect an equipment to this room");
         }
-        return back()->with('error', "Error with the room creation");
+        return back()->with("error", "Error with the room creation");
     }
 
     /**
@@ -49,15 +51,16 @@ class EquipedController extends Controller
      */
     public function edit(Room $room)
     {
+        $this->authorize("update", $room);
         $room_equipement = [];
-        foreach (Equiped::where('room_id', $room->id)->get() as $item) {
-            $room_equipement[] = $item->equipment;
+        foreach (Equiped::where("room_id", $room->id)->get() as $item) {
+            $room_equipement[] = $item;
         }
-        Session::put('room', $room->id);
-        return view('equiped.edit', [
-            'room_equipement' => $room_equipement,
-            'available' => Equipment::where('is_available', true)->get(),
-            'room' => $room,
+        Session::put("room", $room->id);
+        return view("equiped.edit", [
+            "room_equipement" => $room_equipement,
+            "available" => Equipment::where("is_available", true)->get(),
+            "room" => $room,
         ]);
     }
 
@@ -66,28 +69,29 @@ class EquipedController extends Controller
      */
     public function destroy(Equiped $equiped)
     {
-        $equiped->equipment->update(['is_available' => true]);
+        $this->authorize("delete", $equiped);
+        $equiped->equipment->update(["is_available" => true]);
         $equiped->delete();
         return back()->with("success", "You have deleted equipments connected to this room");
     }
 
     public function select(Request $request)
     {
-        $selected_equipment = $request->input('equipment');
+        $selected_equipment = $request->input("equipment");
 
-        if (Session::has('room')) {
-            $room_id = Session::get('room');
-            // Boucle pour ajouter chaque équipement sélectionné à la table equiped
+        if (Session::has("room")) {
+            $room_id = Session::get("room");
+            $this->authorize("update", Room::find($room_id));
+
             foreach ($selected_equipment as $equipment_id) {
                 $equipement = Equipment::find($equipment_id);
-                Equiped::create(['room_id' => $room_id, 'equipment_id' => $equipment_id]);
-                $equipement->update(['is_available' => false]);
+                Equiped::create(["room_id" => $room_id, "equipment_id" => $equipment_id]);
+                $equipement->update(["is_available" => false]);
             }
 
-            // Redirection vers la page de la room avec un message de succès
             return redirect()->route("room.show", ["room" => $room_id])
-                ->with('success', 'The selection of equipment has been successfully registered');
+                ->with("success", "The selection of equipment has been successfully registered.");
         }
-        return back()->with('error', "Error with the room creation");
+        return back()->with("error", "Error with the room creation.");
     }
 }
